# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Finance App API",
        "description": "REST API for personal finance management built with Node.js, Express, Prisma and PostgreSQL.\n\n## Features\n- JWT authentication (access + refresh tokens)\n- User profile under `/api/users/me`\n- Financial transactions with pagination, filters and safe sorting\n- Standardized responses: `{ success, message, data }` and errors `{ success, message, code }`\n- Rate limiting, CORS, Helmet and ownership checks on protected resources\n\n## Authentication\nProtected routes require the header:\n`Authorization: Bearer <accessToken>`\n\nObtain tokens via `POST /api/users` (register) or `POST /api/users/login`. Use `POST /api/users/refresh-token` to renew tokens.\n\n## Notes\n- Passwords are never returned in API responses.\n- `user_id` in request bodies is ignored; the authenticated user from the JWT is always used.\n- Unsupported query parameters (e.g. `categoryId`) return `400 VALIDATION_ERROR`.",
        "version": "1.0.0",
        "contact": {
            "name": "Finance App",
            "url": "https://github.com/vivianeaguiarc/finance-app"
        }
    },
    "host": "finance-app-i600.onrender.com",
    "basePath": "/",
    "schemes": ["https", "http"],
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token. Format: `Authorization: Bearer <accessToken>`"
        }
    },
    "paths": {
...
        "/api/budgets": {
            "post": {
                "tags": ["Budgets"],
                "summary": "Create monthly budget",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "categoryId",
                                "month",
                                "year",
                                "limitAmount"
                            ],
                            "properties": {
                                "categoryId": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "month": { "type": "integer", "example": 6 },
                                "year": { "type": "integer", "example": 2025 },
                                "limitAmount": {
                                    "type": "number",
                                    "example": 500
                                }
                            }
                        }
                    }
                ],
                "responses": { "201": { "description": "Created" } }
            },
            "get": {
                "tags": ["Budgets"],
                "summary": "List monthly budgets",
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "OK" } }
            }
        },
        "/api/budgets/status": {
            "get": {
                "tags": ["Budgets"],
                "summary": "Budget alerts by month",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "$ref": "#/parameters/DashboardMonthParam" },
                    { "$ref": "#/parameters/DashboardYearParam" }
                ],
                "responses": { "200": { "description": "OK" } }
            }
        },
...
    },
    "parameters": {
...
        "DashboardMonthParam": {
            "name": "month",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "description": "Filter by calendar month (1-12). Use with `year` or defaults to current year."
        },
        "DashboardYearParam": {
            "name": "year",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1970,
            "maximum": 2100,
            "description": "Filter by calendar year. Can be combined with `month`."
        },
...
    },
...
}